Move some namespaces on generator
