+++ phpstan.neon
@@ -26,6 +26,7 @@ parameters:
         - %currentWorkingDirectory%/src/Component/Annotation/*
         - %currentWorkingDirectory%/src/Component/Factory/*
         - %currentWorkingDirectory%/src/Component/Metadata/*
+        - %currentWorkingDirectory%/src/Component/Generator/*
         - %currentWorkingDirectory%/src/Component/Reflection/ClassReflection.php
         - %currentWorkingDirectory%/src/Component/spec/*
         - %currentWorkingDirectory%/src/Component/Storage/*
