+++ src/Bundle/Resources/config/services.xml
@@ -37,6 +37,7 @@
         <defaults public="true" />
 
         <service id="sylius.random_generator" class="Sylius\Component\Resource\Generator\RandomnessGenerator" />
+        <service id="Sylius\Resource\Generator\RandomnessGeneratorInterface" alias="sylius.random_generator" />
         <service id="Sylius\Component\Resource\Generator\RandomnessGeneratorInterface" alias="sylius.random_generator" />
 
         <service id="sylius.form.type_extension.form.request_handler"
