+++ src/Component/Generator/RandomnessGenerator.php
@@ -13,54 +13,10 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Resource\Generator;
 
-use Webmozart\Assert\Assert;
+class_exists(\Sylius\Resource\Generator\RandomnessGenerator::class);
 
-final class RandomnessGenerator implements RandomnessGeneratorInterface
-{
-    private string $uriSafeAlphabet;
-
-    private string $digits;
-
-    public function __construct()
+if (false) {
+    final class RandomnessGenerator extends \Sylius\Resource\Generator\RandomnessGenerator
     {
-        $this->digits = implode(range(0, 9));
-
-        $this->uriSafeAlphabet =
-            implode(range(0, 9))
-            . implode(range('a', 'z'))
-            . implode(range('A', 'Z'))
-            . implode(['-', '_', '~'])
-        ;
-    }
-
-    public function generateUriSafeString(int $length): string
-    {
-        return $this->generateStringOfLength($length, $this->uriSafeAlphabet);
-    }
-
-    public function generateNumeric(int $length): string
-    {
-        return $this->generateStringOfLength($length, $this->digits);
-    }
-
-    public function generateInt(int $min, int $max): int
-    {
-        return random_int($min, $max);
-    }
-
-    private function generateStringOfLength(int $length, string $alphabet): string
-    {
-        $alphabetMaxIndex = strlen($alphabet) - 1;
-
-        Assert::greaterThanEq($alphabetMaxIndex, 1);
-
-        $randomString = '';
-
-        for ($i = 0; $i < $length; ++$i) {
-            $index = random_int(0, $alphabetMaxIndex);
-            $randomString .= $alphabet[$index];
-        }
-
-        return $randomString;
     }
 }
